backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
aiofiles==24.1.0
slowapi==0.1.9
